requirements.txt
altair==5.5.0
annotated-types==0.7.0
attrs==25.3.0
autograd==1.8.0
autograd-gamma==0.5.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
et_xmlfile==2.0.0
fonttools==4.58.4
formulaic==1.1.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.173.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.73.0
grpcio-status==1.71.0
httplib2==0.22.0
idna==3.10
interface-meta==1.3.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lifelines==0.30.0
MarkupSafe==3.0.2
matplotlib==3.10.3
narwhals==1.43.1
numpy==2.3.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
proto-plus==1.26.1
protobuf==5.29.5
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rpds-py==0.25.1
rsa==4.9.1
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.46.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
watchdog==6.0.0
wrapt==1.17.2
xlrd==2.0.2
